refactor 24.10.11 syb price 분리
