+++ src/main/java/lotto/README.md
@@ -7,7 +7,9 @@
 
 ### LottoGame
 + [x] 로또 발급
-+ [x] 로또 구입금액 0이하일 경우 에러
-+ [x] 로또 구입금액 게임횟수로 안나눠질 경우 에러
 + [x] 총 수익률 확인
 
+### LottoPrice
++ [x] 로또 구입금액 0이하일 경우 에러
++ [x] 로또 구입금액 게임횟수로 안나눠질 경우 에러
++ [x] 로또 횟수 구하기
